requirements.txt
ibm-watson-machine-learning==1.0.339
langchain==0.0.333
numpy==1.24.2
pandas==1.5.3
pypdf==3.12.2
python-dotenv==1.0.0
requests==2.31.0
rouge==1.0.1
streamlit==1.24.1
safetensors==0.3.1
urllib3==1.26.11
chromadb==0.4.17
pymupdf==1.23.6
ibm-generative-ai==0.5.0